\docType{package}
\name{jpplantnames-package}
\alias{jpplantnames}
\alias{jpplantnames-package}
\title{Unofficial Japanese Plant Name Lookup Tools}
\description{
Fetches and caches the Vascular Plant Japanese Name Checklist ver. 1.10 and
provides helpers for looking up scientific plant names from Japanese names.
}
\details{
This package is not affiliated with or endorsed by JBIF, YList, or the
checklist authors. Checklist data is not included in the package and is not
covered by the package license.
}
