\docType{package}
\name{ylistjp-package}
\alias{ylistjp}
\alias{ylistjp-package}
\title{Unofficial Access to YList Japanese Plant Name Data}
\description{
Fetches and caches the public tab-delimited YList data file and provides
helpers for looking up scientific plant names from Japanese names.
}
\details{
This package is not affiliated with or endorsed by YList. YList data is not
included in the package and is not covered by the package license.
}
